Create Paginate::Paginator::Simple class for paginating Arrays and similar.
This represents a different approach that initially, creating a set of classes actually responsible for the pagination (dubbed, "Paginators". It's the best I could think of). The modules' #paginate methods will delegate the actual work to a Paginator class. The main reason for this approach is using inheritance to minimize code duplication.